feat(front-client): components refacto
